Finished HW3.
Was unable to run the entire code for all FD/SD users, but code works on samples of all types of relevant users.
(For example, I tried to scrape the IDs of SD followers, and I ran it for 20 hours, scraping 36k ids, and it
still wasn't done.)

Several parts of the code are redunant, and could probably be made more parsimonious/elegant, but it took a while
to get it to the point where it works are intended, and due to the rate limits it was difficult to keep editing and
trying different things.